opt_dff: split up
