+++ passes/opt/dff/CMakeLists.txt
@@ -1,0 +1,9 @@
+yosys_pass(opt_dff
+	opt_dff.cc
+	simple.cc
+	constbits.cc
+	eqbits.cc
+	opt_dff.h
+	REQUIRES
+		simplemap
+)
